requirements.txt
matplotlib==2.2.3
prettytable==0.7.2
python-dateutil==2.7.4
requests==2.20.0
